Refactor Enums module to a separate file
